phy_dev/Mi Smart Desk Lamp Lite/result-01/logic_code/requirements.txt
python-miio
